+++ HotWiredBridge/src/hotwiredbridge/wired/UserListEvent.java
@@ -3,13 +3,13 @@ package hotwiredbridge.wired;
 import java.util.List;
 
 public class UserListEvent extends WiredEvent {
-	private int chatId;
+	private long chatId;
 	private List<User> users;
 
-	public int getChatId() {
+	public long getChatId() {
 		return chatId;
 	}
-	public void setChatId(int chatId) {
+	public void setChatId(long chatId) {
 		this.chatId = chatId;
 	}
